Refactoring logging to be at top level (#790)
