+++ shared/src/index.ts
@@ -8,5 +8,5 @@ export { runGenerate } from './commands/generate/generate.js';
 export { ValidationOutput } from './commands/validate/validation.output.js';
 export { CALM_META_SCHEMA_DIRECTORY } from './consts.js';
 export { SchemaDirectory } from './schema-directory.js';
-export { initLogger } from './commands/helper.js';
+export { initLogger } from './logger.js';
 export * from './types';
