+++ shared/src/schema-directory.ts
@@ -3,7 +3,7 @@ import { join } from 'path';
 import pointer from 'json-pointer';
 import { mergeSchemas, updateStringValuesRecursively } from './util.js';
 import { Logger } from 'winston';
-import { initLogger } from './commands/helper.js';
+import { initLogger } from './logger.js';
 
 /**
  * Stores a directory of schemas and resolves references against that directory.
